Update Sensum dataset
